Remove UbuntuDesktopInstaller dependencies from individual pages
WSL OOBE wizard is going to reuse the WelcomePage, but the app will be
different with another set of routes. Providing the app locale helps to
eliminate the dependency to UbuntuDesktopInstaller.

Locale and theme are similar in the way that changing either rebuilds
the whole app, so the proposal is to put them together to avoid yet
another tiny provider.