Remove LocalizedView
`LocalizedView` cannot be shared in `ubuntu_wizard:widgets`, because
`AppLocalizations` is an app-specific type, and Dart does not allow
accessing static members of generic types.

Access `AppLocalizations` directly using the same syntax view models,
`Settings`, `Wizard`, and many built-in Flutter types such as `Theme`
and `MediaQuery` are accessed.

`AppLocalizations` is available via `l10n.dart` and the getter is non-
nulla non-nullable since #246, which makes ot convenient enough to use.

As a bonus, this reduces the widget tree depth by one level in a
majority of pages and widgets.