Create a launcher script to set up the right environment
Since the snap is classic we are not using the standard gnome launcher.
Without the right variables set we use the gdk-pixbuf loaders cache from
the system. In newer Ubuntu series the png loader is built-in but that's
the case for the version bundled with the installer which leads to have no
png loader available and gtk bailing out while trying to load icons.

Fixes #323