Add hidden --initial-route command-line argument for integration testing
This allows integration tests jump straight on a specific page.

For example, if an integration test wants to test navigation to the
"Turn off RST" or "Turn off BitLocker" pages, the test doesn't need to
navigate through half a dozen pages _and_ enter irrelevant input
values on the way to the target page. With help of --initial-route, the
test can jump straight on the page before and test that the wizard
shows the correct page when pressing "Continue".