Resolve non-deterministic code generation issue
It appears that exporting subiquity_client's Disk & Partition types from
view models instead of the disk storage service stabilizes the generated
output from build_runner.

It could be because the disk storage service has to be imported to both
the view and its model. When generating a mock model mock for testing,
the same data class types are visible via multiple import dependencies.