Remove bogus keyboard layout loading from the welcome screen
As documented in KeyboardService.load(), the locale must be passed to
subiquity before loading keyboard layouts. However, when browsing
languages on the welcome page, the locale is not passed to subiquity
until the user presses the Continue button. Therefore, the UI loads
the same en_US keyboard layouts again and again every time the user
changes the UI language.

Loading the keyboard layouts is better done in the keyboard layout
page, where the layouts and variants are actually used. At that point,
the locale has been applied too.

Fixes: #363