Fix Subiquity initialization in WSL
In commit 04c3d4e9, the desktop installer was changed to request geo-IP
detection for the timezone. The problem is that such endpoint doesn't
exist in the WSL system setup, so an exception is thrown on startup:

    Unhandled Exception: setTimezone("geoip") returned error 404

This commit adds a callback for subiquity initialization routines to
make it straight-forward to make any variant-specific initialization
calls on startup, as soon as the connection has been established.