Specify country code in .arb files
Instead of supplying multiple .arb files per language to get a full
locale, construct full locales with help of a country code field
specified in the base .arb file.

This allows us to remove a lot of extra .arb files that were basically
unused, but had to be there to satisfy Flutter (flutter/flutter#92731).