Limit the window size to the available workarea (#542)
When the available workarea is smaller than the default window size,
maximize the window to let the window manager take care of filling the
available workarea. This ensures that the window fits on the screen
even if the resolution is relatively low. For example, the default in
Virtual Box is 800x600.

Ref: #518