WSL: fix language selection in the integration test (#554)
The test was assuming that the "French" list item was visible on the
screen by default. However, #530 added a long list of supported
languages, which pushed the "French" item out of the screen when the
default "English" item is selected on startup. Consequently, the test
would not be able to tap the "French" item. This commit fixes the issue
by scrolling the list until the "French" item becomes visible, before
trying to tap it.