Resolve the location of subiquity_client without symlinks
This change makes use of the `package_config` Dart package to resolve
the location of the `subiquity_client` because the symlinks to
`subiquity_client/subiquity` and `.subiquity/socket` in the project
tree are problematic in many ways.

Not only we've had to duplicate the symlinks for both the desktop
installer and the WSL setup, but any desktop installer flavor would
have to create similar symlinks too, to be able to dry-run the
installer flavor.

Also, the symlinks are problematic for snap builds. The socket symlinks
break local (as in source-type: local) builds. Using Git as the
source-type circumvents this problem, but the drawback is that then
snapcraft doesn't pick local uncommitted changes.

Close: #477