Fix RadioIconTile layout
RadioIconTile was originally added for aligning arbitrary widgets with
RadioListTile. It has been replaced by RadioButton, which has similar
but not identical layout. This change fixes the layout mismatch by
adapting RadioButton's layout to RadioIconTile, and then reusing it for
implementing RadioButton's and CheckButton's layout. This eliminates a
nice chunk of duplication between RadioButton and CheckButton, and
ensures that all three stay in sync.