Fix clunky dropdown focus visuals (#774)
The default focus highlight from the Material design DropdownButton
does not match well with the input decoration that is used to styleour
dropdowns. Remove the default focus color and let the input decorator
highlight focus instead. This way, the focus is consistently
highlighted in the same way as input fields do.