Gives real work to the ViewModel.
Previously the ApplyingChangesModel was just exposing a raw stream.
Now it's doing a business of a real model, which also simplifies the
view a lot.
Also, the framework treats better the call to Wizard.next() coming from
this ChangeNotifier than by reaction to a raw stream.
Even though, just to be extra cautious, the model will only call the
VoidCallback parameter once, at the time of transition.