[subiquity_client] sync IdentityData with Subiquity
All the IdentityData fields are non-nullable i.e. none of them is
Optional in Subiquity. Making every field nullable only gives a false
sense of safety and clutters the UI code with unnecessary null-aware
operators and repeated imaginary defaults that are never used because
the values coming from Subiquity cannot be null.

Ref: #412