Ensures port is used on Uri.http
In my experimentations setting the port on the connection factory is not
enough when Uri.http is used, because it's expected that the
full authority will be specified (i.e. "host:port").
This creates the authority getter on the endpoint to facilitate
formatting uri's.
Also subiquity client, server and monitor are modified to
take this new method in consideration when building HTTP uri's.