Adds TCP Helpers
Useful to safely find a port available on the system.
Subiquity will bind on that port later, on the Linux side.
This socket will be closed before the client attempts to connect.
This way the localhost loopback interface forwarding will just work.
Automagically.
It's required that the port range is agreed between Linux and Windows.