Installation slides page model
This is quite different from the ubuntu_wsl_splash.
Encapsulates visibility of the journal view and
the status of the installation:
   which goes from WSL registering until proceed to setup.
   moment in which the slide show must call Wizard.next.
Must know Subiquity status to let the wizard flow when the server is up.
Errors may raise even before subiquity can start.
Error recovery capability is topic for future changes.