Await window_manager initialization
window_manager basically does nothing without it on Windows OS.
On Linux it does nothing. On Windows it stores the top-level HWND.
The plugin could retrieve it from the view()->GetNativeWindow().
That could be less performant, I guess.
Anyway, it's easier to fix it here than on the plugin and upgrading it.