DiskStorageService: expose guided storage choice (#988)
The information will be used for implementing the navigation flow for
guided storage resize (#819).

In principle, a guided storage target is automatically chosen whenever
possible i.e. if there's a) only one disk to erase or b) enough free
space for the installation. If an automatic choice is made, the
respective pages for selecting or resizing a guided storage are skipped.
The value of this property will be checked in the wizard router when
choosing the next page.