Add "Install Ubuntu alongside Y" page
This is just the page and its route extracted from #992 without changes
to the wizard flow. The possibility to select alongside-installation on
the "Install type" page and the wizard flow changes will follow in the
final step in #992.

The page can be tested by selecting the initial route and passing a
suitable machine config. For example:

```sh
$ flutter run -a "--initial-route=/installalongside" \
              -a "--machine-config=examples/win10-along-ubuntu.json"
```