Fix telemetry report location
Write the report to /var/log instead of /target/var/log because the GUI
has no permission to write to the latter. The report will be copied to
its final place on the target system by the upcoming post-install script.
Furthermore, in debug/dry-run mode, write to the app dir similarly to
the logs.