Replace RadioIconTile with simple padding (#1085)
The reason for using RadioIconTile was to get an indentation that matches
with RadioListTile and RadioButton. Given that RadioIconTile does not
play nicely with AnimatedExpanded and Flexible, the easiest solution is
to define the indentation that matches with kRadioSize + 8px spacing so
that it can be easily used with a plain simple Padding widget.