Post-pone guided storage choice
A guided storage choice consists of multiple attributes:

- which disk/partition to use
- whether to use lvm
- whether to use encryption
- security key

These inputs are collected from multiple pages:

- installation type
- select guided storage
- install x alongside y
- choose security key

The problem is that Subiquity only allows setting guided storage once.
That is, we cannot "update" the existing guided storage selection as we
go but we need to wait until all attributes have been gathered and then
apply it all in one go. The best place for this is when arriving at the
"Write changes to disk" page where the storage changes are summarized.