Skip listening if the server is already DONE
During integration testing it's possible that ApplyingChangesModel start listening after the server is already DONE.
If so, the next state change would be to null (shutdown) which only happens if the GUI ends.
But GUI won't exit because it's expecting a status change.
Infinite loop.
If the assumption is correct.