README.md: update instructions
Running `make install_deps` as non-root is better as it depends on
`gitdeps` which builds the local curtin & probert as the local user
and `aptdeps` that uses sudo to install the relevant dependencies.
The `installdeps.sh` script is somewhat intrusive because it runs
dist-upgrade etc. and is supposedly primarily meant for the CI.