Revise workarea detection & fix window size comparison
First of all, use the default window size (960x680) instead of the size
allocation including shadows (1012x732) for determining whether it fits
on the screen.

Secondly, given that gdk_monitor_get_workarea() is not reliable early
on startup and we have to subtract some arbitrary margins, make sure to
subtract the margins from the full geometry instead of the workarea to
avoid subtracting double if and when the workarea is correctly reported
when restarting the installer after startup.

Fixes: #1500