Allocate disk space: present "leave unformatted" as "none" format
Previously, we were presenting "leave unformatted" as null. However, we
want to have an additional separator which will be supported by the
upcoming MenuButtonBuilder in ubuntu_widgets and presented as null. Thus,
present the "leave unformatted" option with an explicit value to be able
to use null for the separator.