Present fake "install alongside" that leads to "bitlocker instructions"
If Subiquity doesn't offer any "install alongside" option and we detect
BitLocker on the system, assume that the option got filtered out due to
BitLocker. At this point, we're already past the minimum installation
size check so we can safely assume that there's enough disk space in
total. In this case, we want to present a fake "install alongside"
option that leads to BitLocker instructions.