Postpone source & storage calls
The storage call is necessary on startup to fetch such information that
is needed to build the wizard (has rst, total disk space, ...). A valid
installation source must be selected before the storage call, however,
it must only be done in manual installation mode to avoid conflicting
with autoinstall source config.