Manual partitioning: fix initial GUI state update (#1827)
GET storage/v2 and storage/v2/orig_config requests were sent in parallel
but only the former notified the GUI to update. Depending on the order
of completion, the GUI could have left in a wrong initial state that
confusingly stated that existing partitions would be formatted because
it didn't see the original storage config. Reorder the calls to make
sure both configs are available before notifying the GUI to update.