Remove Wizard(Action).done
It was used to finish the telemetry report and write it on the disk but
that approach didn't work because Subiquity copies it to /target when it
finishes installing and it's too late to write anything from the GUI
when it gets closed. Also, writing any telemetry data from early exits
is irrelevant because it doesn't end up stored anywhere anyway.