Make WSL part of the CI aware of pubspec lockfile
We aim with this experiment with committing the pubspec.lock
file into the repository. Since WSL is smaller and final, it's a good use
case to try first.
Currently I'm trying not to restrict to much the SDK version.
That can lead to more noise. At least the CI should create a PR
automatically when divergence appears.